requirements.txt
Django==3.2.5
djangorestframework==3.13.1
django-filter==21.1
psycopg2-binary==2.9.3

